refactor: frmAddUpdatePerson refactoring and validation enhancements
### 🚀 Refactoring & Enhancements (frmAddUpdatePerson & Validation)

#### 1. Form Constructors & Load Logic
- Implemented overloaded constructors: parameterless for AddNew mode and parameterized (int PersonID) for Update mode.
- Moved form initialization and data loading logic from constructors to the Form_Load event handler to eliminate code duplication across constructors.

#### 2. Input Validation & ErrorProvider Bug Fixes
- Extracted validation logic into a dedicated helper class for modularity.
- Enhanced email validation using a custom Regex pattern via a dedicated boolean helper method.
- Bug Fix: Fixed an issue where the ErrorProvider icon persisted when the user cleared the email field after an invalid entry. Resolved by clearing the error (SetError(control, null)) when the field becomes empty.

#### 3. Data Back Delegate Implementation
- Added a custom DataBack delegate/event in frmAddUpdatePerson to return the saved PersonID back to caller forms, adhering to the course design pattern.

+++ DVLD_Buisness/clsPerson.cs
@@ -17,7 +17,7 @@ namespace DVLD_BusinessLogic
         public string ThirdName { get; set; }
         public string LastName { get; set; }
         public DateTime DateOfBirth { get; set; }
-        public byte Gender { get; set; }
+        public short Gender { get; set; }
         public string Address { get; set; }
         public string Phone { get; set; }
         public string Email { get; set; }
